split menu repository for example of query paging util
