+++ src/main/java/com/example/menu/repository/MenuQueryRepository.java
@@ -1,0 +1,9 @@
+package com.example.menu.repository;
+
+import com.example.menu.entity.Menu;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+interface MenuQueryRepository {
+    Page<Menu> findAllByShopId(Pageable pageable, int id);
+}
